Fix bug #11071: modification of rgb node
git-svn-id: https://scm.gforge.inria.fr/svn/openalea/trunk@2819 04ea39c6-a500-0410-8508-e77fe2a861cf

+++ stdlib/src/openalea/color/py_color.py
@@ -43,76 +43,76 @@ color_list=[(0,0,0),
             ]
 
 class ColorNode (Node) :
-	"""Small color node
-	"""
-	def __init__ (self, inputs, outputs) :
-		Node.__init__(self, inputs, outputs)
-		self.set_caption(" ")
-		self.get_ad_hoc_dict().set_metadata("useUserColor",True)
-	
-	def __call__ (self, inputs) :
-		color = list(inputs[0])
-		if len(color) in (3,4) :
-			self.get_ad_hoc_dict().set_metadata("useUserColor",True)
-			self.get_ad_hoc_dict().set_metadata("userColor",color)
-		
-		return color,
+    """Small color node
+    """
+    def __init__ (self, inputs, outputs) :
+        Node.__init__(self, inputs, outputs)
+        self.set_caption(" ")
+        self.get_ad_hoc_dict().set_metadata("useUserColor",True)
+    
+    def __call__ (self, inputs) :
+        color = list(inputs[0])
+        if len(color) in (3,4) :
+            self.get_ad_hoc_dict().set_metadata("useUserColor",True)
+            self.get_ad_hoc_dict().set_metadata("userColor",color)
+        
+        return color,
 
 class FixedColorNode (Node) :
-	"""Small color node
-	"""
-	_color = None
-	def __init__ (self, inputs, outputs) :
-		Node.__init__(self, inputs, outputs)
-		self.set_caption(" ")
-		self.get_ad_hoc_dict().set_metadata("useUserColor",True)
-		self.get_ad_hoc_dict().set_metadata("userColor",self._color)
-	
-	def __call__ (self, inputs) :
-		return self._color,
+    """Small color node
+    """
+    _color = None
+    def __init__ (self, inputs, outputs) :
+        Node.__init__(self, inputs, outputs)
+        self.set_caption(" ")
+        self.get_ad_hoc_dict().set_metadata("useUserColor",True)
+        self.get_ad_hoc_dict().set_metadata("userColor",self._color)
+    
+    def __call__ (self, inputs) :
+        return self._color,
 
 class BlackNode (FixedColorNode) :
-	_color = [0,0,0]
+    _color = [0,0,0]
 
 class WhiteNode (FixedColorNode) :
-	_color = [255,255,255]
+    _color = [255,255,255]
 
 class RedNode (FixedColorNode) :
-	_color = [255,0,0]
+    _color = [255,0,0]
 
 class GreenNode (FixedColorNode) :
-	_color = [0,255,0]
+    _color = [0,255,0]
 
 class BlueNode (FixedColorNode) :
-	_color = [0,0,255]
+    _color = [0,0,255]
 
 def col_item (ind) :
-	return color_list[ind % len(color_list)],
+    return color_list[ind % len(color_list)],
 
 def random (alpha) :
-	col = tuple(rd.random() * 255 for i in range(3) )
-	if alpha is not None :
-		col += (alpha,)
-	
-	return col,
+    col = tuple(rd.random() * 255 for i in range(3) )
+    if alpha is not None :
+        col += (alpha,)
+    
+    return col,
 
 def rgb (H,S,V, alpha) :
-	rgb_col = tuple(v * 255 for v in hsv_to_rgb(H / 360.,S / 255.,V / 255.) )
-	
-	if alpha is not None :
-		rgb_col += (alpha,)
-	
-	return rgb_col,
+    rgb_col = tuple(int(v * 255) for v in hsv_to_rgb(H / 360.,S / 255.,V / 255.) )
+    
+    if alpha is not None :
+        rgb_col += (alpha,)
+    
+    return rgb_col,
 
 def hsv (rgb_col) :
-	r,g,b = (v / 255. for v in rgb_col[:3])
-	h,s,v = rgb_to_hsv(r,g,b)
-	H,S,V = (h * 360,s * 255,v * 255)
-	
-	if len(rgb_col) == 3 :
-		return H,S,V,None
-	else :
-		return H,S,V,rgb_col[3]
+    r,g,b = (v / 255. for v in rgb_col[:3])
+    h,s,v = rgb_to_hsv(r,g,b)
+    H,S,V = (int(h * 360),int(s * 255),int(v * 255))
+    
+    if len(rgb_col) == 3 :
+        return H,S,V,None
+    else :
+        return H,S,V,rgb_col[3]
 
 def color_map(val, minval=0, maxval=1, coul1=80, coul2=20):
     """todo"""
